requirements.txt
asgiref==3.4.1
Django==4.0
django-environ==0.8.1
django-rest-framework==0.1.0
django-simple-history==3.0.0
djangorestframework==3.12.4
sqlparse==0.4.2
